moving functions from internal/trace_utils to internal/trace_utils_base
