+++ ddtrace/contrib/internal/trace_utils.py
@@ -21,17 +21,19 @@ from urllib import parse
 
 import wrapt
 
+from ddtrace.contrib.internal.trace_utils_base import USER_AGENT_PATTERNS  # noqa:F401
+from ddtrace.contrib.internal.trace_utils_base import _get_header_value_case_insensitive
+from ddtrace.contrib.internal.trace_utils_base import _get_request_header_user_agent
 from ddtrace.contrib.internal.trace_utils_base import _normalize_tag_name
+from ddtrace.contrib.internal.trace_utils_base import _set_url_tag
+from ddtrace.contrib.internal.trace_utils_base import set_user  # noqa:F401
 from ddtrace.ext import http
 from ddtrace.ext import net
-from ddtrace.ext import user
 from ddtrace.internal import core
 from ddtrace.internal.compat import ensure_text
 from ddtrace.internal.compat import ip_is_global
 from ddtrace.internal.core.event_hub import dispatch
 from ddtrace.internal.logger import get_logger
-from ddtrace.internal.utils.http import redact_url
-from ddtrace.internal.utils.http import strip_query_string
 import ddtrace.internal.utils.wrappers
 from ddtrace.propagation.http import HTTPPropagator
 from ddtrace.settings._config import config
@@ -59,8 +61,6 @@ RESPONSE = "response"
 # starting a "new object" on the UI.
 NORMALIZE_PATTERN = re.compile(r"([^a-z0-9_\-:/]){1}")
 
-# Possible User Agent header.
-USER_AGENT_PATTERNS = ("http-user-agent", "user-agent")
 
 IP_PATTERNS = (
     "x-forwarded-for",
@@ -73,24 +73,6 @@ IP_PATTERNS = (
     "cf-connecting-ip",
     "cf-connecting-ipv6",
 )
-
-
-def _get_header_value_case_insensitive(headers, keyname):
-    # type: (Mapping[str, str], str) -> Optional[str]
-    """
-    Get a header in a case insensitive way. This function is meant for frameworks
-    like Django < 2.2 that don't store the headers in a case insensitive mapping.
-    """
-    # just in case we are lucky
-    shortcut_value = headers.get(keyname)
-    if shortcut_value is not None:
-        return shortcut_value
-
-    for key, value in headers.items():
-        if key.lower().replace("_", "-") == keyname:
-            return value
-
-    return None
 
 
 def _store_headers(headers, span, integration_config, request_or_response):
@@ -121,23 +103,6 @@ def _store_headers(headers, span, integration_config, request_or_response):
             continue
         # An empty tag defaults to a http.<request or response>.headers.<header name> tag
         span.set_tag_str(tag_name or _normalize_tag_name(request_or_response, header_name), header_value)
-
-
-def _get_request_header_user_agent(headers, headers_are_case_sensitive=False):
-    # type: (Mapping[str, str], bool) -> str
-    """Get user agent from request headers
-    :param headers: A dict of http headers to be stored in the span
-    :type headers: dict or list
-    """
-    for key_pattern in USER_AGENT_PATTERNS:
-        if not headers_are_case_sensitive:
-            user_agent = headers.get(key_pattern)
-        else:
-            user_agent = _get_header_value_case_insensitive(headers, key_pattern)
-
-        if user_agent:
-            return user_agent
-    return ""
 
 
 def _get_request_header_referrer_host(headers, headers_are_case_sensitive=False):
@@ -402,22 +367,6 @@ def ext_service(pin, int_config, default=None):
     return default
 
 
-def _set_url_tag(integration_config, span, url, query):
-    # type: (IntegrationConfig, Span, str, str) -> None
-    if not integration_config.http_tag_query_string:
-        span.set_tag_str(http.URL, strip_query_string(url))
-    elif config._global_query_string_obfuscation_disabled:
-        # TODO(munir): This case exists for backwards compatibility. To remove query strings from URLs,
-        # users should set ``DD_TRACE_HTTP_CLIENT_TAG_QUERY_STRING=False``. This case should be
-        # removed when config.global_query_string_obfuscation_disabled is removed (v3.0).
-        span.set_tag_str(http.URL, url)
-    elif getattr(config._obfuscation_query_string_pattern, "pattern", None) == b"":
-        # obfuscation is disabled when DD_TRACE_OBFUSCATION_QUERY_STRING_REGEXP=""
-        span.set_tag_str(http.URL, strip_query_string(url))
-    else:
-        span.set_tag_str(http.URL, redact_url(url, config._obfuscation_query_string_pattern, query))
-
-
 def set_http_meta(
     span,  # type: Span
     integration_config,  # type: IntegrationConfig
@@ -628,58 +577,6 @@ def set_flattened_tags(
     for prefix, value in items:
         for tag, v in _flatten(value, sep, prefix, exclude_policy):
             span.set_tag(tag, processor(v) if processor is not None else v)
-
-
-def set_user(
-    tracer,  # type: Any
-    user_id,  # type: str
-    name=None,  # type: Optional[str]
-    email=None,  # type: Optional[str]
-    scope=None,  # type: Optional[str]
-    role=None,  # type: Optional[str]
-    session_id=None,  # type: Optional[str]
-    propagate=False,  # type bool
-    span=None,  # type: Optional[Span]
-    may_block=True,  # type: bool
-    mode="sdk",  # type: str
-):
-    # type: (...) -> None
-    """Set user tags.
-    https://docs.datadoghq.com/logs/log_configuration/attributes_naming_convention/#user-related-attributes
-    https://docs.datadoghq.com/security_platform/application_security/setup_and_configure/?tab=set_tag&code-lang=python
-    """
-    if span is None:
-        span = core.get_root_span()
-    if span:
-        if user_id:
-            str_user_id = str(user_id)
-            span.set_tag_str(user.ID, str_user_id)
-            if propagate:
-                span.context.dd_user_id = str_user_id
-
-        # All other fields are optional
-        if name:
-            span.set_tag_str(user.NAME, name)
-        if email:
-            span.set_tag_str(user.EMAIL, email)
-        if scope:
-            span.set_tag_str(user.SCOPE, scope)
-        if role:
-            span.set_tag_str(user.ROLE, role)
-        if session_id:
-            span.set_tag_str(user.SESSION_ID, session_id)
-
-        if (may_block or mode == "auto") and asm_config._asm_enabled:
-            exc = core.dispatch_with_results("set_user_for_asm", [tracer, user_id, mode]).block_user.exception
-            if exc:
-                raise exc
-
-    else:
-        log.warning(
-            "No root span in the current execution. Skipping set_user tags. "
-            "See https://docs.datadoghq.com/security_platform/application_security/setup_and_configure/"
-            "?tab=set_user&code-lang=python for more information.",
-        )
 
 
 def extract_netloc_and_query_info_from_url(url):
